Move TimerQueue.TickCount Windows implementation to shared CoreLib
