+++ src/System.Private.CoreLib/System.Private.CoreLib.csproj
@@ -276,9 +276,6 @@
     <Compile Include="shared\Interop\Windows\Kernel32\Interop.HandleTypes.cs" />
     <Compile Include="shared\Interop\Windows\Kernel32\Interop.GetStdHandle.cs" />
     <Compile Include="shared\Interop\Windows\Kernel32\Interop.LocalAlloc.cs" />
-    <Compile Include="shared\Interop\Windows\Kernel32\Interop.QueryUnbiasedInterruptTime.cs" />
-    <Compile Include="shared\Interop\Windows\Kernel32\Interop.VerifyVersionExW.cs" />
-    <Compile Include="shared\Interop\Windows\Kernel32\Interop.VerSetConditionMask.cs" />
     <Compile Include="shared\Interop\Windows\Ole32\Interop.CoTaskMemAlloc.cs" />
     <Compile Include="shared\Interop\Windows\OleAut32\Interop.SysAllocStringByteLen.cs" />
   </ItemGroup>
