+++ src/System.Private.CoreLib/shared/System/Environment.WinRT.cs
@@ -13,6 +13,8 @@ namespace System
 
         public static string UserDomainName => "Windows Domain";
 
+        internal static readonly bool IsWindows8OrAbove = true;
+
         private static string GetFolderPathCore(SpecialFolder folder, SpecialFolderOption option)
         {
             WinRTInteropCallbacks callbacks = WinRTInterop.UnsafeCallbacks;
