+++ src/System.Private.CoreLib/shared/System/Threading/TimerQueue.Windows.cs
@@ -1,0 +1,40 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.InteropServices;
+
+namespace System.Threading
+{
+    internal partial class TimerQueue
+    {
+        private static int TickCount
+        {
+            get
+            {
+                // We need to keep our notion of time synchronized with the calls to SleepEx that drive
+                // the underlying native timer.  In Win8, SleepEx does not count the time the machine spends
+                // sleeping/hibernating.  Environment.TickCount (GetTickCount) *does* count that time,
+                // so we will get out of sync with SleepEx if we use that method.
+                //
+                // So, on Win8, we use QueryUnbiasedInterruptTime instead; this does not count time spent
+                // in sleep/hibernate mode.
+                if (Environment.IsWindows8OrAbove)
+                {
+                    ulong time100ns;
+
+                    bool result = Interop.Kernel32.QueryUnbiasedInterruptTime(out time100ns);
+                    if (!result)
+                        Marshal.ThrowExceptionForHR(Marshal.GetLastWin32Error());
+
+                    // convert to 100ns to milliseconds, and truncate to 32 bits.
+                    return (int)(uint)(time100ns / 10000);
+                }
+                else
+                {
+                    return Environment.TickCount;
+                }
+            }
+        }
+    }
+}
